chore: unify duplicate ClaimRecord definitions
