+++ examples/multi_agent_human_chat/agents/claims/dspy_modules/claims_data.py
@@ -1,10 +1,9 @@
 """Claims data store and tools for claim operations."""
 
 import json
-import re
 from typing import Dict, List, Optional, Tuple, Union
 
-from pydantic import BaseModel, Field, field_validator
+from agents.claims.types import ClaimRecord
 
 from libraries.logger import get_console_logger
 from libraries.tracing import create_tracer
@@ -12,64 +11,6 @@ from libraries.tracing.otel import safe_set_attribute
 
 from .claims_errors import ErrorCategory, ErrorResponse, get_user_friendly_error
 
-logger = get_console_logger("claims_agent.data")
-tracer = create_tracer("claims_agent_data")
-
-
-class ClaimRecord(BaseModel):
-    """Data structure for an insurance claim record with validation."""
-
-    claim_number: str = Field(..., description="Unique identifier for the claim")
-    policy_number: str = Field(
-        ..., description="Policy number associated with the claim"
-    )
-    status: str = Field(..., description="Current status of the claim")
-    next_steps: str = Field(..., description="Next steps required for claim processing")
-    outstanding_items: List[str] = Field(
-        default_factory=list, description="Items pending for claim processing"
-    )
-    estimate: Optional[float] = Field(None, description="Estimated payout amount", gt=0)
-    estimate_date: Optional[str] = Field(
-        None, description="Estimated date for payout (YYYY-MM-DD)"
-    )
-    details: Optional[str] = Field(
-        None, description="Detailed description of the claim"
-    )
-    address: Optional[str] = Field(None, description="Address related to the claim")
-    phone: Optional[str] = Field(None, description="Contact phone number")
-    damage_description: Optional[str] = Field(
-        None, description="Description of damage or loss"
-    )
-    contact_email: Optional[str] = Field(None, description="Contact email address")
-
-    @field_validator("estimate_date")
-    @classmethod
-    def validate_date_format(cls, v):
-        """Validate date is in YYYY-MM-DD format."""
-        if v is None:
-            return v
-        if not re.match(r"^\d{4}-\d{2}-\d{2}$", v):
-            raise ValueError("Date must be in YYYY-MM-DD format")
-        return v
-
-    @field_validator("phone")
-    @classmethod
-    def validate_phone(cls, v):
-        """Validate phone number format."""
-        if v is None:
-            return v
-        # Simple validation - could be enhanced with country-specific patterns
-        if not re.match(r"^\+?[\d\-\(\) ]{7,}$", v):
-            raise ValueError("Invalid phone number format")
-        return v
-
-    def to_dict(self) -> Dict:
-        """Convert claim record to dictionary, excluding None values."""
-        return {k: v for k, v in self.model_dump().items() if v is not None}
-
-    def to_json(self) -> str:
-        """Convert claim record to JSON string."""
-        return json.dumps(self.to_dict())
 
 
 # Sample in-memory claims database
